Add Linux support: CMake build, /proc backend, privileged helper, bundling
TaskExplorer was Windows-only. This adds a Linux port alongside it, driven by
the same CMake description intended to build both.

Build
- Replace the qmake .pro/.pri files with CMake. The .vcxproj files remain as
  the reference for what the Windows build contained. The Windows build has
  not been compiled from CMake yet.

Linux backend (TaskExplorer/API/Linux)
- Processes, threads, handles, modules, memory maps and windows from /proc,
  sock_diag netlink for sockets, systemd over D-Bus for daemons, /proc/modules
  and /sys/module for drivers.
- Views with no Windows counterpart: control groups, LSM/seccomp security,
  pressure stall information.
- ELF core dumps in the same format the kernel and gcore produce, verified
  loadable in gdb with full backtraces.
- X11 window enumeration, including processes that set no _NET_WM_PID, via the
  X-Resource extension.

TaskHelper
- The existing IPC helper now builds on Linux as a Qt-free worker, using the
  same length-prefixed CVariant protocol as the Windows one. It is spawned
  elevated on demand for the things that need privilege: /proc reads refused
  to this user, open file descriptors, memory reads, libdw stack traces, and
  the thread-stopping half of a core dump.
- Off by default behind Tasks -> Use Privileged Helper, because enabling it
  raises an authentication prompt, and a task manager must not do that merely
  because it was started.
- Every exit path releases a process stopped for a dump, including SIGKILL.
- debuginfod is disabled before touching libdwfl; it otherwise blocks
  indefinitely fetching debug info over the network for stripped modules.

Packaging
- Qt and the libraries that cannot be relied upon are bundled next to the
  executable and found through $ORIGIN, so the output directory is
  self-contained. Installer/README-portable.md covers why this needs DT_RPATH
  rather than DT_RUNPATH, and why the glibc floor is decided by the build host
  rather than by what gets bundled.

Also fixes a pre-existing bug in shared code: the dump progress dialog was
connected to OnProgressMessage/OnStatusMessage, which do not exist on
CProgressDialog - the slots are ShowProgress/ShowStatus. Progress never
updated and dump failures were never shown at all.

+++ MiscHelpers/Common/SmartGridWidget.cpp
@@ -48,7 +48,15 @@ void CSmartGridWidget::ReArange()
 			count++;
 	}
 
-	float columns = ceil(sqrt((float)count));
+	//
+	// double rather than float deliberately: the float overload resolves to
+	// sqrtf, and glibc 2.43 introduced a new symbol version of that function.
+	// Linking against it raised the binary's minimum glibc from 2.34 to 2.43 -
+	// i.e. from "most distributions of the last few years" to "only the newest"
+	// - for one square root of a small integer. sqrt(double) has carried the
+	// same GLIBC_2.2.5 version since forever.
+	//
+	float columns = ceil(sqrt((double)count));
 	float rows = columns > 0 ? ceil(count / columns) : 0;
 
 	for (int row = 0; row < rows; row++)
